feat: Add document validation tool and enhance search server documentation
